Moved popup to settings_dialog
However, settings_dialog doesn't close properly...
