+++ data/settings.ini
@@ -1,7 +1,7 @@
 [general]
 displaySettings: no
 displayPrint: no
-printAllTemplates: yes
+printAllTemplates: no
 repeat: no
 
 [letterGeneration]
